added my notes

+++ staff/blanca-manrique/back-end/notapp/logic/src/deleteNote.1.js
@@ -1,20 +1,16 @@
-const { models: { User, Note }} = require('data')
+const { models: { Note } } = require('data')
+const { validators: { validateId } } = require('commons')
 
 function deleteNote(userId, noteId) {
-    return User.findById(userId) 
-        .then(user => {
-            if (!user) throw new Error(`user with id ${userId} not found`)
+    validateId(userId, 'user id')
+    validateId(noteId, 'note id')
 
-            return Note.findById(noteId)
+    return Note.deleteOne({ user: userId, _id: noteId })
+        .then(result => {
+            const { deletedCount } = result
+
+            if (deletedCount === 0) throw new Error(`note with id ${noteId} and user id ${userId} not found`)
         })
-        .then(note => {
-            if (!note) throw new Error(`note with id ${noteId} not found`)
-
-            if (note.user.toString() !== userId) throw new Error(`note with id ${noteId} does not correspond to user with id ${userId}`)
-
-            return Note.deleteOne({ _id: noteId })
-        })
-        .then(result => {})
 }
 
 module.exports = deleteNote
